feat: redirect messages
Check if sender.id is in the database and redirect to destination.
On every message all the tracked chats are being loaded directly from
the database.
I'm not sure if it's a good way to do it or is it better to have a
local copy which gets updated periodically from the database.
I'll do some research and will eventually change it if needed.