fix: improve id parse
Found out that a chat id can be negative
for groups. So I decided to modify the part of the code
which parses the id.
Also found out that it sometimes doesn't work for different
types of chats. For example it works fine if a message is redirected
from a channel to a channel but it sometimes doesn't work if
a message is redirected from a channel to a group.

Closes #2.